# api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API para registro e consulta de contratos de financiamento automotivo utilizando blockchain Hyperledger Besu e contratos inteligentes UUPS",
        "title": "MobX API - VFinance Registry",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MobX API Support",
            "email": "support@mobx-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "3.0.0"
    },
    "host": "144.22.179.183:3000",
    "basePath": "/",
    "paths": {
        "/api/auth/token": {
            "post": {
                "description": "Gera um token JWT válido para autenticação",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Generate JWT Token",
                "responses": {
                    "200": {
                        "description": "Token gerado com sucesso",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "valid": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/validate": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Valida um token JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate JWT Token",
                "responses": {
                    "200": {
                        "description": "Token válido",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "exp": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "role": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "user_id": {
                                    "type": "string"
                                },
                                "valid": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou ausente",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                },
                                "valid": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Digite \"Bearer\" seguido de um espaço e o token JWT",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}